requirements.txt
uwsgi
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
coreapi==2.3.3
markdown==3.3.6
django-filter==21.1
django-guardian==2.4.0
Pillow==9.0.1
requests==2.27.1
selenium==4.1.3
cn2an==0.5.16
beautifulsoup4==4.10.0
requests[socks]
